backend/pyproject.toml
[project]
name = "documents-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "PyJWT>=2.9.0",
    "cryptography>=43.0.0",
    "pyyaml>=6.0",
    "openai>=1.50.0",
]
